+++ x/posts/internal/keeper/common_test.go
@@ -27,7 +27,7 @@ func SetupTestInput() (sdk.Context, keeper.Keeper) {
 		panic(err)
 	}
 
-	// create a Cdc and a context
+	// create a cdc and a context
 	cdc := testCodec()
 	ctx := sdk.NewContext(ms, abci.Header{ChainID: "test-chain-id"}, false, log.NewNopLogger())
 
@@ -65,7 +65,7 @@ var testPost = types.NewPost(
 	testPostCreationDate,
 	testPostOwner,
 ).WithMedias(types.NewPostMedias(
-	types.NewPostMedia("https://uri.com", "text/plain"),
+	types.NewPostMedia("https://uri.com", "text/plain", []sdk.AccAddress{testPostOwner}),
 )).WithPollData(types.NewPollData(
 	"poll?",
 	testPostEndPollDate,