+++ x/posts/internal/types/models/post_response.go
@@ -4,17 +4,15 @@ import (
 	"encoding/json"
 	"fmt"
 	"strings"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
 // PostQueryResponse represents the data of a post
 // that is returned to user upon a query
 type PostQueryResponse struct {
 	Post
-	PollAnswers []UserAnswer            `json:"poll_answers,omitempty" yaml:"poll_answers,omitempty"`
-	Reactions   []ReactionQueryResponse `json:"reactions" yaml:"reactions,omitempty"`
-	Children    PostIDs                 `json:"children" yaml:"children"`
+	PollAnswers []UserAnswer   `json:"poll_answers,omitempty" yaml:"poll_answers,omitempty"`
+	Reactions   []PostReaction `json:"reactions" yaml:"reactions,omitempty"`
+	Children    PostIDs        `json:"children" yaml:"children"`
 }
 
 // String implements fmt.Stringer
@@ -28,7 +26,7 @@ Children: %s
 }
 
 func NewPostResponse(
-	post Post, pollAnswers []UserAnswer, reactions []ReactionQueryResponse, children PostIDs,
+	post Post, pollAnswers []UserAnswer, reactions []PostReaction, children PostIDs,
 ) PostQueryResponse {
 	return PostQueryResponse{
 		Post:        post,
@@ -56,18 +54,4 @@ func (response *PostQueryResponse) UnmarshalJSON(data []byte) error {
 
 	*response = PostQueryResponse(temp)
 	return nil
-}
-
-type ReactionQueryResponse struct {
-	Value string         `json:"value" yaml:"value"`
-	Code  string         `json:"shortcode" yaml:"shortcode"`
-	Owner sdk.AccAddress `json:"owner" yaml:"owner"`
-}
-
-func NewReactionQueryResponse(value, code string, owner sdk.AccAddress) ReactionQueryResponse {
-	return ReactionQueryResponse{
-		Value: value,
-		Code:  code,
-		Owner: owner,
-	}
 }