+++ x/profile/legacy/v0.6.0/types.go
@@ -1,0 +1,5 @@
+package v060
+
+const (
+	ModuleName = "profiles"
+)